+++ NEWS.md
@@ -1,8 +1,10 @@
 # pathfindR (development version)
 
 ## Major Changes
-- in `run_pathfindR()`, added the argument `create_HTML_report` to make HTML report optional
+- refactored/simplified `run_pathfindR()`
+- visualization enriched term diagrams are now NOT part of `run_pathfindR()`
 - in `hierarchical_term_clustering()`, update the sequence of number of clusters for which silhouette width is calculated for choosing the optimal number of clusters. This should speed up the function for cases with a large number of enriched terms
+
 
 ## Minor Changes
 - fixed minor issue in `return_pin_path()` (#157)